refactoring, small changes
